refector: useEmailInput, usePasswordInput의 중복을 제거하고 useInput 훅으로 공통화
