+++ src/components/NavBar.tsx
@@ -3,11 +3,11 @@ import { css } from '@emotion/react';
 import { FiArrowLeft, FiUser } from 'react-icons/fi';
 import { useNavigate } from 'react-router-dom';
 import theme from '../styles/theme';
-import { PascalCase } from '../pages/Login/UserManagement';
+import { UserManagement } from '../pages/Login/UserManagement';
 
 const NavBar = () => {
   const navigate = useNavigate();
-  const { user } = PascalCase();
+  const { user } = UserManagement();
 
   const goToBack = () => {
     navigate(-1);
